% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDatalist.R
\name{addDatalist}
\alias{addDatalist}
\title{Add Datalist}
\usage{
addDatalist(file = paste0(PATH.STAN, "data\\\\stanDim.rda"),
  list = error("'list' must be specified"))
}
\arguments{
\item{file}{an R data file.}

\item{list}{a list of objects contained in \sQuote{file}.}
}
\description{
Add R data to package
}
\details{
Adding R data files to R packages requires a data list. This is a helper function of \code{\link{loadDim}}.
}
\examples{
x <- c(1:10)
save(x, file = "data.rda")
addDatalist(file = "data.rda", list = c("x"))
}
\author{
OECD STAN
}
\seealso{
\code{\link{packageData}}
}
\keyword{package}

