% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bdHMM-class}
\alias{.bdHMM}
\alias{bdHMM-class}
\alias{[,bdHMM,ANY,ANY,ANY-method}
\title{This class is a generic container for bidirectional Hidden Markov Models.}
\description{
This class is a generic container for bidirectional Hidden Markov Models.
}
\section{Slots}{

\describe{
\item{\code{initProb}}{Initial state probabilities.}

\item{\code{transMat}}{Transition probabilities}

\item{\code{emission}}{Emission parameters as an HMMEmission object.}

\item{\code{nStates}}{Number of states.}

\item{\code{status}}{of the HMM. On of c('initial', 'EM').}

\item{\code{stateNames}}{State names.}

\item{\code{dimNames}}{Names of data tracks.}

\item{\code{LogLik}}{Log likelihood of a fitted HMM.}

\item{\code{transitionsOptim}}{There are three methods to choose from for fitting the transitions. Bidirectional transition matrices (invariant under reversal of time and direction) can be fitted using c('rsolnp', 'ipopt'). 'None' uses standard update formulas and the resulting matrix is not constrained to be bidirectional.}

\item{\code{directedObs}}{An integer indicating which dimensions are directed. Undirected dimensions are 0. Directed observations must be marked as unique integer pairs. For instance c(0,0,0,0,0,1,1,2,2,3,3) contains 5 undirected observations, and thre pairs (one for each direction) of directed observations.}

\item{\code{dirScore}}{Directionlity score of states of a fitted bdHMM.}
}}
\section{Methods}{
  \describe{
    \item{[}{get elements from the bdHMM}
    }
}
\examples{
nStates = 5
stateNames = c('F1', 'F2', 'R1', 'R2', 'U1')
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
myEmission = list(d1=HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means)))

bdhmm = bdHMM(initProb=initProb, transMat=transMat, emission=myEmission, nStates=nStates, status='initial', stateNames=stateNames, transitionsOptim='none', directedObs=as.integer(0))

}
\seealso{
\code{\linkS4class{HMMEmission}}
}

