% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{bdhmm_ex}
\alias{bdhmm_ex}
\title{Initial bdHMM for the Quick-Start example in the vignette}
\description{
Initial bdHMM for the Quick-Start example in the vignette
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

