% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdHMMTransMatOptim.R
\name{c2optimize}
\alias{c2optimize}
\title{Optimize transitions}
\usage{
c2optimize(pars)
}
\arguments{
\item{pars}{Parameters for optimization.}
}
\value{
optimized transitions
}
\description{
The function is called from C++ to optimize transitions.
}

