% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFormula.R
\name{calcFormula}
\alias{calcFormula}
\title{Calc formula}
\usage{
calcFormula(data = stop("'data' must be specified"),
  formula = stop("'formula' must be specified"), id.vars = c("cou", "var",
  "year"))
}
\arguments{
\item{data}{data in wide format with industries in columns.}

\item{formula}{data frame with "ind" and "formula" column.}

\item{id.vars}{character vector to specify preserved columns.}
}
\description{
Apply industry conversion based on calculation formula
}
\details{
Uses a data frame containing formulas to convert industry classification.
}
\examples{
require(reshape2)
require(stanData)
data(STANNAi4)
cou <- "AUT"
var <- "VALU"
data <- DATA.STAN
data <- data[data$cou\%in\%cou &
             data$var\%in\%var,]
data <- dcast(data, cou + var + year ~ ind, value.var = "value")
file <- paste0(PATH.SASi4, "Lists\\\\MDL_STAN_i4_i3_A64.txt")
line.remove <- c("/* aggregate */")
formula <- SASformula(file=file, line.remove=line.remove)
data <- calcFormula(data=data, formula=formula, id.vars=id.vars)
data <- melt(data, id.vars = id.vars, variable.name = "ind", na.rm = TRUE)
data <- subset(data, select = c("cou", "var", "ind", "year", "value"))
}
\author{
OECD STAN
}
\seealso{
\code{\link{SASformula}}
}
\keyword{SAS}

