% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizePoilog.R
\name{call_dpoilog}
\alias{call_dpoilog}
\title{Calculate density of the Poisson-Log-Normal distribution.}
\usage{
call_dpoilog(x)
}
\arguments{
\item{x}{A vector c(n, mu, sigma), where n is the number of observed counts, mu the mean of the Log-Normal distribution and sigma its variance.}
}
\value{
Density of the Poisson-Log-Normal distribution.
}
\description{
Calculate density of the Poisson-Log-Normal distribution.
}
\examples{
call_dpoilog(c(5, 2, 1))

}

