% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convCreateDiag.R
\name{convCreateDiag}
\alias{convCreateDiag}
\title{Create Diagnoal Converter}
\usage{
convCreateDiag(dim = list(row = c(58, 37), col = c(58)), agg.row1 = c(1:10),
  agg.row2 = c(1:18))
}
\arguments{
\item{data}{.}

\item{toRow}{.}

\item{toCol}{.}

\item{displayConv}{.}
}
\description{
Create Diagonal Matrix Aggregation Converters
}
\details{
Create diagonal converters for the aggregation of matrices. Tested with two cascaded row dimensions and one column dimension.
}
\examples{
convCreateDiag(dim=list(row=c(10,5),col=c(10)),agg.row1=c(1:10),agg.row2=c(1:2))
}
\author{
OECD STAN
}
\seealso{
\code{\link{convShow}}
}
\keyword{IO}

