% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convCreateDiag2.R
\name{convCreateDiag2}
\alias{convCreateDiag2}
\title{Create Diagnoal Converter}
\usage{
convCreateDiag2(dim = list(row = c(62, 34), col = c(34)),
  agg.row1 = c(1:62), agg.row2 = c(1:34))
}
\arguments{
\item{data}{.}

\item{toRow}{.}

\item{toCol}{.}

\item{displayConv}{.}
}
\description{
Create Diagonal Matrix Aggregation Converters
}
\details{
Create diagonal converters for the aggregation of matrices. Tested with two cascaded row dimensions and one column dimension.
}
\examples{
convCreateDiag2(dim=list(row=c(10,5),col=c(10)),agg.row1=c(1:10),agg.row2=c(1:2))
}
\author{
OECD STAN
}
\seealso{
\code{\link{convShow}}
}
\keyword{IO}

