% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCurrency.R
\name{convertCurrency}
\alias{convertCurrency}
\title{Convert Currencies}
\usage{
convertCurrency(data = stop("'data' must be specified"),
  datacur = DATA.XRATES[DATA.XRATES$var == "EXCH", ], tounit = "USD")
}
\arguments{
\item{data}{a data frame with dimensions \code{cou}, \code{var}, \code{year} and \code{value}.}

\item{datacur}{a data frame with dimension \code{cou}, \code{var} and \code{value}.}
}
\description{
Apply currency conversion to monetary variables
}
\details{
This function converts monetary values from national currency to US dollars.
}
\examples{
convertCurrency(data=DATA.EUNAMAR2, datacur=DATA.XRATES[DATA.XRATES$var=="EXCH",])
}
\author{
OECD STAN
}
\seealso{
\code{\link{packageData}}
}
\keyword{currency}

