% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convShow.R
\name{convShow}
\alias{convShow}
\title{Show Converter}
\usage{
convShow(m, xLabels = NULL, yLabels = NULL)
}
\arguments{
\item{m}{a matrix.}

\item{xLabels}{labels for x dimension.}

\item{yLabels}{labels for y dimension.}
}
\description{
Show Matrix Aggregation Converters
}
\details{
Show converters for the aggregation of matrices.
}
\examples{
see "convCreate.svg" in /vignettes/diagrams
m <- convCreate(dim=list(row=c(10,5),col=c(10)),agg.row1=c(1:10),agg.row2=c(1:2),agg.col1=c(5),horiz=FALSE,dimnames=NULL)
convShow(m)
}
\author{
OECD STAN
}
\seealso{
\code{\link{convCreate}}
}
\keyword{IO}

