% Generated by roxygen2 (4.0.2): do not edit by hand
\name{couReport}
\alias{couReport}
\title{COU Report}
\usage{
couReport(cou.sou = stop("'cou.sou' must be specified"), isic = 4,
  create.data = TRUE, create.plot = TRUE, namevar.plot = NULL,
  file.brew = file.path(PATH.COUi4, "STAN_collect_load.brew"),
  file.out.prefix = "Report_", date = 2013)
}
\arguments{
\item{cou.sou}{a 3-digit country ISO code character string and source ID.}

\item{isic}{an integer specifying the version of ISIC.}

\item{create.data}{logical controls creation of  SAS input data.}

\item{create.plot}{logical controls creation of data plot with all sources.}

\item{namevar.plot}{a character vector specifying the variables to be plotted.}

\item{file.brew}{a character string specifying the brew file to be used.}

\item{file.out.prefix}{a character string implicitly specifying the output location and filename.}

\item{date}{a character string specifying the collection period in the output filename.}

\item{brew.file}{the name of the brew file located a character string prepending the output filename.}
}
\description{
Create National Source Report
}
\details{
Creates HTML report for a flat data file.
}
\examples{
couReport(cou="FRA")

namecou <- c('BEL', 'DEU', 'FIN', 'FRA', 'ITA', 'NLD')
seeds <- as.character(namecou)
path <- paste0(sub("Progs", "Data", PATH.COUi3))
results <- sapply(seeds, couReport,
                  brew.file = paste0(path, "STAN_collect_load.brew"),
                  path.out = paste0(path, cou, "\\\\Rawdata\\\\Report_"))
for (url in results) browseURL(url)
}
\author{
OECD STAN
}
\seealso{
\code{\link{NSO2XLS}}
}
\keyword{SAS}

