% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data2Gviz}
\alias{data2Gviz}
\title{Convert data for plotting with Gviz}
\usage{
data2Gviz(obs, regions, binSize, gen, col = "black")
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{regions}{GRanges object of the regions (e.g. chromosomes) stored in the viterbi path.}

\item{binSize}{The bin size of the viterbi path.}

\item{gen}{The geome id, e.g. hg19, hg38 for human.}

\item{col}{The color of the data tracks.}
}
\value{
A list containing the data tracks converted to Gviz objects for plotting.
}
\description{
Convert data for plotting with Gviz
}

