% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dimMerge}
\alias{dimMerge}
\title{Dim Merge}
\usage{
dimMerge(data, dim, dim1, dim2)
}
\arguments{
\item{data}{a dataframe with "value" dimension and at lease one of "var" or "ind".}

\item{dim}{a character string specifying "var" or "ind".}

\item{dim1}{a character string specifying the first dimension member (x).}

\item{dim2}{a character string specifying the second dimension member (y).}
}
\description{
Dimension Merge
}
\details{
Self-join data frame on two members of same dimension.
}
\examples{
x <- data.frame("ind" = c(rep('D01T03', 2), rep('D05T09', 2)), "var" = rep(c('PROD', 'VALU'), 2), "value" = c(1:4))
x <- dimMerge(data=x, dim="var", dim1="PROD", dim2="VALU")
x <- dimMerge(data=x, dim="ind", dim1="D01T03", dim2="D05T09")
}
\author{
OECD STAN
}
\seealso{
\code{\link{mergeCalc}}
}
\keyword{transform}

