% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dotstatGetData}
\alias{dotStatGetDimensionList}
\alias{dotStatGetDimensionMemberList}
\alias{dotstatGetData}
\title{Access .Stat SQL}
\usage{
dotstatGetData(channel, datasetcode, dim.list, time)

dotStatGetDimensionList(channel, datasetcode, lang)

dotStatGetDimensionMemberList(dim, channel, datasetcode, lang)
}
\arguments{
\item{channel}{open SQL channel created with \code{odbcDriverConnect}}

\item{datasetcode}{character code of dataset}

\item{dim.list}{named list with character vector for each dimension in dataset}

\item{time}{integer vector with start and end year}

\item{lang}{character \code{en} or \code{fr}}

\item{dim}{character to specify dimension}
}
\description{
Use stored procedures to retrieve from OECD.Stat

\code{dotStatGetDimensionList}: Retrieve dimensions of dataset.

\code{dotStatGetDimensionMemberList}: Retrieve dimension members of dataset.
}
\details{
Functions to obtain dimensions, dimension members and data from OECD.Stat datasets

dotStatGetDimensionList(channel, datasetcode, lang)

dotStatGetDimensionMemberList(dim, channel, datasetcode, lang)
}
\examples{
## create the dimension list with members - order of dimensions as in dataset
dim.list.test <- list(LOCATION=c("AUT", "BEL"),
                      VAR=c("PROD", "VALU"),
                      ISIC4=c("05_09", "10_33"),
                      SRC=c("SSIS"),
                      ## SRC=c(""),
                      SCL=c("TOTAL"))
## retrieve data according to selected dimension member selection
data.test <- dotstatGetData(channel = SQL.STAT,
                            datasetcode = "SSIS_BSC_ISIC4",
                            dim.list = dim.list.test,
                            time = c(2007, 2008))
namedim <- dotStatGetDimensionList(channel = SQL.STAT,
                                   datasetcode = "SSIS_BSC_ISIC4",
                                   lang = "en")
## for a single dimension
dotStatGetDimensionMemberList(dim = "LOCATION",
                              channel = SQL.STAT,
                              datasetcode = "SSIS_BSC_ISIC4",
                              lang = "en")
## for multiple dimensions
namedim <- c("LOCATION", "VAR")
dim.list <- lapply(namedim, dotStatGetDimensionMemberList,
                   channel = SQL.STAT,
                   datasetcode = "SSIS_BSC_ISIC4",
                   lang = "en")
names(dim.list) <- namedim
}
\author{
OECD STAN
}
\keyword{SDBS}

