% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{example}
\alias{example}
\title{The data for the bdHMM example in the vignette and examples in the manual}
\description{
The data for the bdHMM example in the vignette and examples in the manual
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

