% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{flags}
\alias{flags}
\title{Pre-computed flag sequence for the Quick-Start example in the vignette}
\description{
Pre-computed flag sequence for the Quick-Start example in the vignette
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

