% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAvgSignal}
\alias{getAvgSignal}
\title{Compute average signal in state segmentation}
\usage{
getAvgSignal(viterbi, obs, fct=mean)
}
\arguments{
\item{viterbi}{A list containing the viterbi paths as factors. The output from getViterbi.}

\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{fct}{The averaging function, default: mean.}
}
\value{
A state x data track matrix containing the average signal.
}
\description{
Compute average signal in state segmentation
}
\examples{
data(yeastTF_databychrom_ex)
nStates = 6
dirobs = as.integer(c(rep(0,10), 1, 1))
bdhmm_gauss = initBdHMM(yeastTF_databychrom_ex, nStates, "Gaussian", directedObs=dirobs)
bdhmm_fitted_gauss = fitHMM(yeastTF_databychrom_ex, bdhmm_gauss)
viterbi_bdhmm_gauss = getViterbi(bdhmm_fitted_gauss, yeastTF_databychrom_ex)
avg_signal = getAvgSignal(viterbi_bdhmm_gauss, yeastTF_databychrom_ex)

}

