% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogLik.R
\name{getLogLik}
\alias{getLogLik}
\title{Calculate log likelihood state distribution.}
\usage{
getLogLik(hmm, obs = list(), emissionProbs = list(), dirFlags = list(), verbose = FALSE, nCores = 1, sizeFactors=matrix(1, nrow=length(obs), ncol=ncol(obs[[1]])))#'
}
\arguments{
\item{hmm}{The Hidden Markov Model.}

\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{emissionProbs}{List of precalculated emission probabilities of emission function is of type 'null'.}

\item{dirFlags}{The flag sequence is needed when a bdHMM is fitted on undirected data (e.g.) ChIP only. It is a \code{list} of \code{character} vectors indication for each position its knwon directionality. U allows all states. F allows undirected states and states in forward direction. R allows undirected states and states in reverse direction.}

\item{verbose}{\code{logical} for printing algorithm status or not.}

\item{nCores}{Number of cores to use for computations.}

\item{sizeFactors}{Library size factors for Emissions PoissonLogNormal or NegativeBinomial as a length(obs) x ncol(obs[[1]]) matrix.}
}
\value{
The log likelihood of the observations sequences, given the model.
}
\description{
The function calculates log likelihood for one or more observation sequence.
}
\examples{

data(example)
hmm_ex = initHMM(observations, nStates=3, method="Gaussian") 
hmm_fitted = fitHMM(observations, hmm_ex)
loglik = getLogLik(hmm_fitted, observations)
loglik

}
\seealso{
\code{\linkS4class{HMM}}
}

