% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSizeFactors}
\alias{getSizeFactors}
\title{Compute size factors}
\usage{
getSizeFactors(obs, celltypes)
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{celltypes}{Indicates the cell type/tissue for each entry in obs.}
}
\value{
A celltype/tissue x data tracks matrix containing the size factors.
}
\description{
Compute size factors
}
\examples{

data(trainRegions)
celltypes = list("E123"=grep("E123", names(trainRegions)), 
        "E116"=grep("E116", names(trainRegions)))
sizeFactors = getSizeFactors(trainRegions, celltypes)
sizeFactors

}

