% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getViterbi.R
\name{getViterbi}
\alias{getViterbi}
\title{Calculate the most likely state path}
\usage{
getViterbi(hmm, obs=list(), NAtol=5, emissionProbs=list(), verbose=FALSE, sizeFactors=matrix(1, nrow=length(obs), ncol=ncol(obs[[1]])))
}
\arguments{
\item{hmm}{The initial Hidden Markov Model.}

\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{NAtol}{Successive positions having NAs longer than this threshold are masked in the viterbi path.}

\item{emissionProbs}{List of precalculated emission probabilities of emission function is of type 'null'.}

\item{verbose}{\code{logical} for printing algorithm status or not.}

\item{sizeFactors}{Library size factors for Emissions PoissonLogNormal or NegativeBinomial as a length(obs) x ncol(obs[[1]]) matrix.}
}
\value{
A list containint the vterbi paths.
}
\description{
Given a Hidden Markov Model, the function calculates the most likely state path (viterbi) for one or more observation sequence.
}
\examples{

data(example)
hmm_ex = initHMM(observations, nStates=3, method="Gaussian") 
hmm_fitted = fitHMM(observations, hmm_ex)
viterbi = getViterbi(hmm_fitted, observations)

}

