% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{HMM-class}
\alias{.HMM}
\alias{HMM-class}
\title{This class is a generic container for Hidden Markov Models.}
\description{
This class is a generic container for Hidden Markov Models.
}
\section{Slots}{

\describe{
\item{\code{initProb}}{Initial state probabilities.}

\item{\code{transMat}}{Transition probabilities}

\item{\code{emission}}{Emission parameters as an HMMEmission object.}

\item{\code{nStates}}{Number of states.}

\item{\code{status}}{of the HMM. On of c('initial', 'EM').}
}}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
HMM(initProb=initProb, transMat=transMat, emission=HMMEmission(type='Gaussian', parameters=list(mean=means, cov=Sigma), nStates=length(means)), nStates=nStates, status='initial')
}
\seealso{
\code{\linkS4class{HMMEmission}}
}

