% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hmm_ex}
\alias{hmm_ex}
\title{Initial HMM for the Quick-Start example in the vignette}
\description{
Initial HMM for the Quick-Start example in the vignette
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

