% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{HMMEmission-class}
\alias{.HMMEmission}
\alias{HMMEmission-class}
\title{This class is a generic container for different emission functions of Hidden Markov Models.}
\description{
This class is a generic container for different emission functions of Hidden Markov Models.
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of emission function c('Gaussian').}

\item{\code{parameters}}{A list containing the the parameters for each state.}

\item{\code{dim}}{Number of dimensions.}

\item{\code{nStates}}{The number of states.}
}}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
HMMEmission(type='Gaussian', parameters=list(mean=means, cov=Sigma), nStates=length(means))
}

