% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HMMEmission}
\alias{HMMEmission}
\title{Create a HMMEmission object}
\usage{
HMMEmission(type = character(), parameters = list(), nStates = numeric())
}
\arguments{
\item{type}{The type of emission function c('Gaussian').}

\item{parameters}{A list containing the the parameters for each state.}

\item{nStates}{The number of states.}
}
\description{
This function creates a HMMEmission object.
}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means))
}

