% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{humanCD4T_flags_ex}
\alias{humanCD4T_flags_ex}
\title{Pre-defined flag sequence for the human CD4T-cell example}
\description{
Pre-defined flag sequence for the human CD4T-cell example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

