% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{humanCD4T_ideogramChr7}
\alias{humanCD4T_ideogramChr7}
\title{Ideogram track of human chromosome 7 for the human CD4T-cell example}
\description{
Ideogram track of human chromosome 7 for the human CD4T-cell example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

