% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{humanCD4T_initMeans}
\alias{humanCD4T_initMeans}
\title{Pre-computed initial estimates of the means for the human CD4T-cell example}
\description{
Pre-computed initial estimates of the means for the human CD4T-cell example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

