% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{humanCD4T_probeAnno_ex}
\alias{humanCD4T_probeAnno_ex}
\title{Genomic positions of processed signal for the human CD4T-cell example}
\description{
Genomic positions of processed signal for the human CD4T-cell example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

