% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{humanCD4T_ucscGenes}
\alias{humanCD4T_ucscGenes}
\title{UCSC gene annotation for the human CD4T-cell example}
\description{
UCSC gene annotation for the human CD4T-cell example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

