% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcIndexVolumeFisher}
\alias{calcIndexPypFisher}
\alias{calcIndexVolumeFisher}
\alias{calcPricePyp}
\alias{calcPypVolume}
\alias{calcVolumePrice}
\alias{calcVolumePyp}
\alias{cpIdxClGeom}
\alias{cpIdxPypGeom}
\alias{cpVolPyp}
\title{calcIndexVolumeFisher}
\usage{
calcIndexVolumeFisher(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  index.var = stop("\\"index.var\\" must be specified"),
  refyear = stop("\\"refyear\\" must be specified"))

calcIndexPypFisher(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  index.var = stop("\\"index.var\\" must be specified"),
  refyear = stop("\\"refyear\\" must be specified"))

calcVolumePyp(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  volume.var = stop("\\"volume.var\\" must be specified"))

calcPypVolume(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  pyp.var = stop("\\"pyp.var\\" must be specified"),
  refyear = stop("\\"refyear\\" must be specified"))

calcVolumePrice(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  volume.var = stop("\\"volume.var\\" must be specified"))

calcPricePyp(data = stop("\\"data\\" must be specified"),
  value.var = stop("\\"value.var\\" must be specified"),
  price.var = stop("\\"price.var\\" must be specified"))

cpIdxClGeom(data = stop("'data' must be specified"), var.cp = "VALU",
  var.idx = "VKOT", id.vars = "ind", refyear = 2005)

cpVolPyp(data = stop("'data' must be specified"), var.cp = "VALU",
  var.cl = "VALK", id.vars = "ind")

cpIdxPypGeom(data = stop("'data' must be specified"), var.cp = "VALU",
  var.idx = "VKOT", id.vars = "ind", refyear = 2005)
}
\arguments{
\item{data}{a dataframe with STAN variables in columns.}

\item{value.var}{STAN variable containing the nominal or quantity series, e.g. \code{VALU}}

\item{index.var}{STAN variable containing the chained index series, e.g. \code{VKOT}}

\item{refyear}{integer to specify the reference year of the chained index series, e.g. \code{2005}}

\item{volume.var}{STAN variable containing the volume series, e.g. \code{VALK}}

\item{pyp.var}{STAN variable containing the previous-year price series, e.g. \code{VKPY}}

\item{price.var}{STAN variable containing the price series, e.g. \code{VALP}}

\item{var.cp}{STAN variable containing the nominal or quantity series, e.g. \code{VALU}}

\item{var.idx}{STAN variable containing the chained index series, e.g. \code{VKOT}}

\item{data}{a dataframe with STAN variables in columns.}

\item{price.var}{STAN variable containing the price series, e.g. \code{VALP}}

\item{pyp.var}{STAN variable containing the previous-year price series, e.g. \code{VKPY}}

\item{volume.var}{STAN variable containing the volume series, e.g. \code{VALK}}

\item{refyear}{integer to specify the reference year of the chained index series, e.g. \code{2005}}
}
\description{
\code{calcIndexVolumeFisher}: calculate volume series using nominal series and Fisher chained index series

\code{calcIndexPypFisher}: calculate previous year price series using nominal series and Fisher chained index series

\code{calcVolumePyp}: calculate previous year price series using nominal series and volume series

\code{calcPypVolume}: calculate volume series using nominal series and previous year price series

\code{calcVolumePrice}: calculate price series using nominal series and volume series

\code{calcPricePyp}: calculate previous year price series using nominal series and price series

\code{cpIdxClGeom}: calculate volume series using nominal series and Fisher chained index series

\code{cpVolPyp}: calculate previous year price series using nominal series and Fisher chained index series

\code{cpIdxPypGeom}: calculate previous year price series using nominal series and Fisher chained index series
}
\details{
calcIndexVolumeFisher(data, value.var, index.var, refyear)

calcIndexPypFisher(data, value.var, index.var, refyear)

calcVolumePyp(data, value.var, index.var, refyear)

calcPypVolume(data, value.var, index.var, refyear)

calcVolumePrice(data, value.var, index.var, refyear)

calcPricePyp(data, value.var, index.var, refyear)

cpIdxClGeom(data, value.var, var.idx, refyear)

cpVolPyp(data, var.cp, var.cl, id.vars)

cpIdxPypGeom(data, var.cp, var.idx, id.vars, refyear)
}
\examples{
data.stan.d <- read.csv(file = system.file(file.path("extdata", "volumeFisherPivot.csv"), package = "stan"))
data.stan.d$VALK2 <- calcIndexVolumeFisher(data = data.stan.d, value.var="VALU", index.var="VKOT", refyear=2005)
data.stan.d$VALK2 <- calcPypVolume(data = data.stan.d, value.var="VALU", pyp.var="VKPY2", refyear=2005)
data.stan.d$VALP2 <- calcVolumePrice(data = data.stan.d, value.var="VALU", volume.var="VALK")
data.stan.d$VKPY2 <- calcPricePyp(data = data.stan.d, value.var="VALU", price.var="VALP")
data.stan.d$VKPY2 <- calcVolumePyp(data = data.stan.d, value.var="VALU", volume.var="VALK2")
data.stan.d <- read.csv(file = system.file(file.path("extdata", "volumeFisherPivot.csv"), package = "stan"))
data.stan.d$VALK <- cpIdxCl(data=data.stan.d,
                            var.cp="VALU",
                            var.idx="VKOT",
                            id.vars="ind",
                            refyear=2005)
var.trans <- data.frame(var.cp  = c("VALU", "PROD", "INTI", "GFCF", "CAPN", "CAPG"),
                        var.idx = c("VKOT", "PKOT", "IKOT", "GKOT", "CKOT", "CGOT"),
                        var.cl  = c("VALK", "PRDK", "INTK", "GFCK", "CPNK", "CPGK"),
                        var.pyp = c("VKPY", "PKPY", "IKPY", "GKPY", "CNPY", "CGPY"))
}
\author{
OECD STAN

OECD STAN
}
\seealso{
\code{\link{estimate}}

\code{\link{estimate}}
}
\keyword{chaining}

