% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initHMM.R
\name{initHMM}
\alias{initHMM}
\title{Initialization of hidden Markov models}
\usage{
initHMM(obs, nStates, method, sizeFactors = matrix(1, nrow = length(obs), ncol = ncol(obs[[1]])), sharedCov = FALSE)
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{nStates}{The number of states.}

\item{method}{Emission distribution of the model. One out of c("NegativeBinomial", "PoissonLogNormal", "NegativeMultinomial", "ZINegativeBinomial", "Poisson", "Bernoulli", "Gaussian", "IndependentGaussian")}

\item{sizeFactors}{Library size factors for Emissions PoissonLogNormal or NegativeBinomial as a length(obs) x ncol(obs[[1]]) matrix.}

\item{sharedCov}{If TRUE, (co-)variance of (Independent)Gaussian is shared over states. Only applicable to 'Gaussian' or 'IndependentGaussian' emissions. Default: FALSE.}
}
\value{
A HMM object.
}
\description{
Initialization of hidden Markov models
}
\examples{

data(example)
hmm_ex = initHMM(observations, nStates=3, method="Gaussian") 

}

