% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{InitProb}
\alias{InitProb}
\title{Get initial state probabilities of a (bd)HMM}
\usage{
InitProb(hmm)
}
\arguments{
\item{hmm}{An object of class HMM or bdHMM.}
}
\value{
The initial state probabilities as a numeric vector.
}
\description{
This function returns the initial state probabilities of a (bd)HMM.
}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
hmm = HMM(initProb=initProb, transMat=transMat, emission=HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means)), nStates=nStates, status='initial')
InitProb(hmm)

}
\seealso{
\code{\linkS4class{HMM}, \linkS4class{bdHMM}}
}

