% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2rmd.R
\name{list2rmd}
\alias{list2rmd}
\title{Read Metadata}
\usage{
list2rmd(list = stop("'list' must be specified"), label = NULL,
  formula = NULL, text.prepend = NULL, print = FALSE, file = NULL)
}
\arguments{
\item{list}{nested list with two levels}

\item{label}{list containing labels for names of 'list'}

\item{formula}{list containing formulas for name of 'list'}

\item{print}{print code for list creation to console}

\item{file}{the exported \code{rdata} file.}
}
\description{
Read metadata from xls saved as tab delimited txt file
}
\details{
Create list with non-empty metadata from file
}
\examples{
## see /data-raw/create_vignette/stanIndicator.R
file <- file.path(dbpath, "GitHub", "jekyll", "industry", "data_stan_notes.rmd")
}
\author{
OECD STAN
}
\seealso{
\code{\link{packageData}}
}
\keyword{metadata}

