% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{LogLik}
\alias{LogLik}
\title{Get stateNames of a (bd)HMM}
\usage{
LogLik(hmm)
}
\arguments{
\item{hmm}{An object of class HMM or bdHMM.}
}
\value{
Log likelihood during model fitting.
}
\description{
This function returns the Log-Likelihood of a (bd)HMM.
}
\examples{
data(example)
hmm_ex = initHMM(observations, nStates=3, method="Gaussian")
hmm_fitted = fitHMM(observations, hmm_ex)
LogLik(hmm_fitted)

}

