% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{observations}
\alias{observations}
\title{Observation sequence for the 'example' data.}
\description{
Observation sequence for the 'example' data.
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

