% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{pilot.hg19}
\alias{pilot.hg19}
\title{Genomic positions of processed signal for the Roadmap Epigenomics data set. Regions from the ENCODE pilot phase.}
\description{
Genomic positions of processed signal for the Roadmap Epigenomics data set. Regions from the ENCODE pilot phase.
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

