% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetadata.R
\name{readMetadata}
\alias{readMetadata}
\title{Read Metadata}
\usage{
readMetadata(file = stop("'file' must be provided"), level = "lower",
  print = FALSE)
}
\arguments{
\item{file}{the exported \code{rdata} file.}

\item{level}{type of metadata: \code{higher} and \code{lower}}

\item{print}{print code for list creation to console}
}
\description{
Read metadata from xls saved as tab delimited txt file
}
\details{
Create list with non-empty metadata from file
}
\examples{
## see /data-raw/create_vignette/stanIndicator.R
file <- system.file("extdata", "stanIndic_meta_lower.txt", package = "stan")
}
\author{
OECD STAN
}
\seealso{
\code{\link{packageData}}
}
\keyword{metadata}

