% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runningMean}
\alias{runningMean}
\title{Smooth data with running mean}
\usage{
runningMean(x, winHalfSize = 2)
}
\arguments{
\item{x}{A vector with the data.}

\item{winHalfSize}{The smoothing window half size.}
}
\value{
A vector containing the smoothed data.
}
\description{
Smooth data with running mean
}
\examples{

data(trainRegions)
celltypes = list("E123"=grep("E123", names(trainRegions)), 
        "E116"=grep("E116", names(trainRegions)))
sizeFactors = getSizeFactors(trainRegions, celltypes)
sizeFactors

}

