% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAScouCreate.R
\name{SAScouCreate}
\alias{SAScouCreate}
\title{SAS COU Create}
\usage{
SAScouCreate(cou = stop("'cou' must be specified"), isic = 3,
  period = "2013-14", remove = FALSE)
}
\arguments{
\item{cou}{a 3-digit country ISO code character string.}

\item{isic}{an integer to specify the version of STAN.}

\item{period}{a character string with update period.}

\item{remove}{logical remove country from STAN SAS tables (country, units).}
}
\description{
Create SAS Country
}
\details{
Sets up the SAS folder structure for a new country in STAN.
}
\examples{
couCreate(cou="BGR", isic=3)
}
\author{
OECD STAN
}
\seealso{
\code{\link{SAScouList}}, \code{\link{SASbatch}}
}
\keyword{SAS}

