% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SASpub.R
\name{SASpub}
\alias{SASpub}
\title{SAS pub}
\usage{
SASpub(cou = stop("'cou' must be specified"), isic = 3, pub.ind = NULL,
  del.ind = NULL, quiet = FALSE)
}
\arguments{
\item{cou}{a 3-digit country ISO code character string.}

\item{isic}{an integer to specify the version of STAN.}

\item{pub.ind}{a character vector restricting the industries to publish.}

\item{del.ind}{a character vector naming the industries to remove before publication.}

\item{quiet}{a logical to return the created SAS program contents.}
}
\description{
Create SAS publication programs
}
\details{
Create or modify SAS publication programs for STAN based on a list of countries and sources.
}
\examples{
SASpub(cou="BGR", sources=list(BGR = c('NSONA', 'SUT', 'ISPATCH')))
}
\author{
OECD STAN
}
\seealso{
\code{\link{SASbatch}}, , \code{\link{SASload}}, \code{\link{SASestim}}
}
\keyword{SAS}

