% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-package.R
\docType{package}
\name{stan-package}
\alias{stan-package}
\title{The stan Package}
\description{
Structural Analysis.
}
\details{
\tabular{ll}{
Package: \tab stan\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-09-10\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://www.oecd.org/sti/stan\cr
}

Provides functions for Structural Analysis
}
\author{
STAN OECD
}
\keyword{package}

