% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stanVolume.R
\name{stanVolume}
\alias{cpIdxCl}
\alias{cpIdxPyp}
\alias{cpVolPyp}
\alias{stanVolume}
\title{stanVolume}
\usage{
cpIdxCl(data = stop("'data' must be specified"), var.cp = "VALU",
  var.idx = "VKOT", id.vars = "ind", refyear = 2010)

cpVolPyp(data = stop("'data' must be specified"), var.cp = "VALU",
  var.cl = "VALK", id.vars = "ind")

cpIdxPyp(data = stop("'data' must be specified"), var.cp = "VALU",
  var.idx = "VKOT", id.vars = "ind", refyear = 2010)
}
\arguments{
\item{data}{a dataframe with STAN variables in columns.}

\item{var.cp}{STAN variable containing the nominal or quantity series, e.g. \code{VALU}}

\item{var.idx}{STAN variable containing the chained index series, e.g. \code{VKOT}}

\item{refyear}{integer to specify the reference year of the chained index series, e.g. \code{2005}}

\item{price.var}{STAN variable containing the price series, e.g. \code{VALP}}

\item{pyp.var}{STAN variable containing the previous-year price series, e.g. \code{VKPY}}

\item{volume.var}{STAN variable containing the volume series, e.g. \code{VALK}}
}
\description{
\code{cpIdxClGeom}: calculate volume series using nominal series and Laspeyres or Fisher chained index series

\code{cpVolPyp}: calculate previous year price series using nominal series and Laspeyres or Fisher chained volume series

\code{cpIdxPypGeom}: calculate previous year price series using nominal series and Laspeyres or Fisher chained index series
}
\details{
cpIdxClGeom(data, value.var, var.idx, refyear)

cpVolPyp(data, var.cp, var.cl, id.vars)

cpIdxPypGeom(data, var.cp, var.idx, id.vars, refyear)
}
\examples{
data.stan.d <- read.csv(file = system.file(file.path("extdata", "volumeFisherPivot.csv"), package = "stan"))
data.stan.d$VALK <- cpIdxCl(data=data.stan.d,
                            var.cp="VALU",
                            var.idx="VKOT",
                            id.vars="ind",
                            refyear=2005
                            )
var.trans <- data.frame(var.cp  = c("VALU", "PROD", "INTI", "GFCF", "CAPN", "CAPG"),
                        var.idx = c("VKOT", "PKOT", "IKOT", "GKOT", "CKOT", "CGOT"),
                        var.cl  = c("VALK", "PRDK", "INTK", "GFCK", "CPNK", "CPGK"),
                        var.pyp = c("VKPY", "PKPY", "IKPY", "GKPY", "CNPY", "CGPY"))
}
\author{
OECD STAN
}
\seealso{
\code{\link{estimate}}
}
\keyword{Volume}

