% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{StateNames}
\alias{StateNames}
\title{Get stateNames of a (bd)HMM}
\usage{
StateNames(hmm)
}
\arguments{
\item{hmm}{An object of class HMM or bdHMM.}
}
\value{
A character vector
}
\description{
This function returns the names of states.
}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
hmm = HMM(stateNames=as.character(1:5), initProb=initProb, transMat=transMat, emission=HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means)), nStates=nStates, status='initial')
StateNames(hmm)

}

