% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{[,bdHMM,ANY,ANY,ANY-method}
\alias{[,bdHMM,ANY,ANY,ANY-method}
\title{extract parts of bdHMM}
\usage{
\S4method{[}{bdHMM,ANY,ANY,ANY}(x, i, j, ..., drop = "missing")
}
\arguments{
\item{x}{A bidirectional hidden Markov model.}

\item{i}{State ids to extract.}

\item{j}{Emissions to extract.}

\item{...}{...}

\item{drop}{...}
}
\value{
Extract parts of bdHMM
}
\description{
extract parts of bdHMM
}
