% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{[,HMM,ANY,ANY,ANY-method}
\alias{[,HMM,ANY,ANY,ANY-method}
\title{extract parts of HMM}
\usage{
\S4method{[}{HMM,ANY,ANY,ANY}(x, i, j, ..., drop = "missing")
}
\arguments{
\item{x}{A hidden Markov model.}

\item{i}{State ids to extract.}

\item{j}{Emissions to extract.}

\item{...}{...}

\item{drop}{...}
}
\value{
extract parts of HMM
}
\description{
extract parts of HMM
}
