% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{trainRegions}
\alias{trainRegions}
\title{Training regions for the Roadmap Epigenomics data set. Three ENCODE pilot regions with data from two cell lines.}
\description{
Training regions for the Roadmap Epigenomics data set. Three ENCODE pilot regions with data from two cell lines.
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

