% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{ucscGenes}
\alias{ucscGenes}
\title{UCSC gene annotation for the Roadmap Epigenomics data set.}
\description{
UCSC gene annotation for the Roadmap Epigenomics data set.
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

