% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viterbi2GRanges}
\alias{viterbi2GRanges}
\title{Convert the viterbi path to a GRanges object}
\usage{
viterbi2GRanges(viterbi, regions, binSize)
}
\arguments{
\item{viterbi}{A list containing the viterbi paths as factors. The output from getViterbi.}

\item{regions}{GRanges object of the regions (e.g. chromosomes) stored in the viterbi path.}

\item{binSize}{The bin size of the viterbi path.}
}
\value{
The viterbi path as GRanges object.
}
\description{
Convert the viterbi path to a GRanges object
}
\examples{
library(GenomicRanges) 
data(yeastTF_databychrom_ex)
nStates = 6
dirobs = as.integer(c(rep(0,10), 1, 1))
bdhmm_gauss = initBdHMM(yeastTF_databychrom_ex, nStates, "Gaussian", directedObs=dirobs)
bdhmm_fitted_gauss = fitHMM(yeastTF_databychrom_ex, bdhmm_gauss)
viterbi_bdhmm_gauss = getViterbi(bdhmm_fitted_gauss, yeastTF_databychrom_ex)
yeastGRanges = GRanges(IRanges(start=1214616, end=1225008), seqnames="chrIV")
names(viterbi_bdhmm_gauss) = "chrIV"
viterbi_bdhmm_gauss_gr = viterbi2GRanges(viterbi_bdhmm_gauss, yeastGRanges, 8)

}

