% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viterbi2Gviz}
\alias{viterbi2Gviz}
\title{Convert state segmentation for plotting with Gviz}
\usage{
viterbi2Gviz(viterbi, chrom, gen, from, to, statecols)
}
\arguments{
\item{viterbi}{A list containing the viterbi paths as factors. The output from getViterbi.}

\item{chrom}{The chromosome/sequence if to convert.}

\item{gen}{The geome id, e.g. hg19, hg38 for human.}

\item{from}{Genomic start poistion.}

\item{to}{Genomic end poistion.}

\item{statecols}{Named vector with state colors.}
}
\value{
A list containing the viterbi path converted to Gviz objects for plotting.
}
\description{
Convert state segmentation for plotting with Gviz
}

