% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLtransform.R
\name{XMLtransform}
\alias{XMLtransform}
\title{XML transform (EU)}
\usage{
XMLtransform(cou = stop("'cou' must be specified"), isic = 4,
  append = FALSE, SDMX = FALSE)
}
\arguments{
\item{cou}{a character string 3-digit ISO code specifying the country.}

\item{isic}{industry classification to be used for the data output locacation.}

\item{append}{a boolean expression if variables shall be added to those already present.}

\item{SDMX}{use SDMX industry list.}
}
\description{
XML Questionnaire transformation (EU)
}
\details{
Transform EU SDMX Questionnaire for loading in SAS.
}
\examples{
XMLtransform(cou = "POL", isic = 4, append = FALSE, SDMX = TRUE)
}
\author{
OECD STAN
}
\seealso{
\code{\link{ANA2XLS}}
}
\keyword{XML}

