% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{yeastTF_databychrom_ex}
\alias{yeastTF_databychrom_ex}
\title{Processed ChIP-on-chip data for yeast TF example}
\description{
Processed ChIP-on-chip data for yeast TF example
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

