% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{yeastTF_initMeans}
\alias{yeastTF_initMeans}
\title{Pre-computed initial estimates of the means for yeast TF example}
\description{
Pre-computed initial estimates of the means for yeast TF example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

