% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{yeastTF_probeAnno_ex}
\alias{yeastTF_probeAnno_ex}
\title{Genomic positions of the ChIP-on-chip data for yeast TF example}
\description{
Genomic positions of the ChIP-on-chip data for yeast TF example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

