% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{yeastTF_SGDGenes}
\alias{yeastTF_SGDGenes}
\title{SGD annotation for the yeast TF example}
\description{
SGD annotation for the yeast TF example
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

