% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{adjust}
\usage{
adjust(data = stop("'data' must be specified"),
  hierarchy = STANi3.HIERARCHY, top = "CTOTAL")
}
\arguments{
\item{data}{a dataframe object.}

\item{hierarchy}{the hierarchy for top-down adjustment applied to \code{variable.name}.}

\item{top}{highest aggregate in applied \code{hierarchy}}
}
\description{
Adjust time series
}
\details{
Adjust time series using industry hierarchy
}
\examples{
data table: cou, var, ind, year, value
adjust(data=data.ext.start, variable.name="ind", value.var="value", hierarchy=STANi3.HIERARCHY, top="CTOTAL")
}
\author{
OECD STAN
}
\seealso{
\code{\link{extend}}, \code{\link{detail}}
}
\keyword{estimate}

