% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarizeData}
\alias{binarizeData}
\title{Binarize Sequencing data with the default ChromHMM binarization}
\usage{
binarizeData(obs)
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}
}
\value{
Binarized observation sequences as a list.
}
\description{
Binarize Sequencing data with the default ChromHMM binarization
}
\examples{

data(trainRegions)
binData = binarizeData(trainRegions)

}

