% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convCreate.R
\name{convCreate}
\alias{convCreate}
\title{Create Converter}
\usage{
convCreate(dim = list(row = c(58, 37), col = c(58)), agg.row1 = c(1:10),
  agg.row2 = c(1:2), agg.col1 = c(5), horiz = FALSE, dimnames = NULL)
}
\arguments{
\item{dim}{a list with two non-empty elements specifying the dimensions of the converter.}

\item{agg.row1}{a numeric vector with two elements specifying the first row aggregation dimension.}

\item{agg.row2}{a numeric vector with two elements specifying the second row aggregation dimension.}

\item{agg.col1}{a numeric vector with two elements specifying the column aggregation dimension.}

\item{horiz}{logical specifying if the converter shall be applied horizontally.}

\item{dimnames}{a list with as many items as dimensions, each item as many elements as dimension members.}
}
\description{
Create Matrix Aggregation Converters
}
\details{
Create converters for the aggregation of matrices. Tested with two cascaded row dimensions and one column dimension.
}
\examples{
see "convCreate.svg" in /vignettes/diagrams
convCreate(dim=list(row=c(10,5),col=c(10)),agg.row1=c(1:10),agg.row2=c(1:2),agg.col1=c(5),horiz=FALSE,dimnames=NULL)
}
\author{
OECD STAN
}
\seealso{
\code{\link{convShow}}
}
\keyword{IO}

