% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detail.R
\name{detail}
\alias{detail}
\title{detail}
\usage{
detail(data, namesou, ind.parent, ind.peers)
}
\arguments{
\item{data}{a dataframe object.}

\item{namesou}{a character vector specifying the main source and secondary source(s).}
}
\description{
Add detailed shares
}
\details{
Add industry detail to time series using shares from secondary sources
}
\examples{
data table: cou, var, ind, year, value
detail(data=data.det.start, namesou=namesou.det, ind.parent=ind.parent[1], ind.peers=ind.peers)
}
\author{
OECD STAN
}
\seealso{
\code{\link{extend}}
}
\keyword{estimate}

