% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(data = data.all, sources = est.array.d[est.array.d$est == "DET", ],
  period = c(1995:2010), isic = 3)
}
\arguments{
\item{data}{a dataframe object}

\item{sources}{a data frame with main source and secondary sources for each combination in the data}

\item{period}{time period for estimation}

\item{isic}{industry classification}
}
\description{
Create STAN estimates
}
\details{
Combine extend and detail functions to time series according to supplied list of sources
}
\examples{
data table: cou, var, ind, year, value
sources table: cou, var, ind, est, MAIN, SEC
data.all <- stani3Estimate.data.all # source stani3Estimate.R
exportcou <- c("NOR")
exportvar <- c("VALU", "PROD")
nameind <- as.character(ui.stani3Estimate.ind)
nameest <- c("DET", "EXT")
nameyear <- c(1995:2012)
est.array <- ui.stani3Estimate.est.array[exportcou, exportvar, nameind, nameest,,drop = FALSE]
est.array.m <- melt(est.array, id.vars = c("cou", "var", "ind", "est"), variable.name = "sou")
est.array.m$ind <- factor(est.array.m$ind, levels = STANi3.INDA60All)
est.array.d <- dcast(est.array.m, cou + var + ind + est ~ sou, value.var = "value")
x <- match(unique(est.array.m$ind), names(STANi3.HIERARCHYINV))
x <- x[!is.na(x)]
ind.parent <- unique(unlist(STANi3.HIERARCHYINV[x]))
data.all <- data.all[data.all$cou\%in\%as.character(unique(est.array.m$cou)) &
                     data.all$var\%in\%unique(est.array.m$var) &
                     data.all$ind\%in\%union(unique(est.array.m$ind), ind.parent) &
                     data.all$year\%in\%nameyear &
                     data.all$sou\%in\%unique(unlist(strsplit(as.character(est.array.m$value), split = ", "))),]

res.ext <- NULL
data.ext <- data.all
print(est.array.d[est.array.d$est=="EXT",])
res.ext <- estimate(data=data.ext,
                    sources=est.array.d[est.array.d$est=="EXT",],
                    period=nameyear)
## add back rows from main source
sou.main.ext <- melt(ui.stani3Estimate.est.array[exportcou, exportvar, nameind, "EXT", "MAIN"])
names(sou.main.ext) <- sub("value", "sou", names(sou.main.ext))
data.all.ext.main <- merge(data.all, sou.main.ext)
res.ext <- rbind(res.ext, data.all.ext.main) # adds around 6000 rows from data.all
res.ext <- res.ext[!duplicated(res.ext[,!colnames(res.ext)\%in\%c("sou", "value")]),]
res.ext$sou <- "PATCHEXT"
res.ext <- rbind(res.ext, data.all)
##
res.det <- NULL
data.det <- res.ext
print(est.array.d[est.array.d$est=="DET",])
res.det <- estimate(data=data.det
                    ,
                    sources=est.array.d[est.array.d$est=="DET",]
                    ,
                    period=nameyear
                    )

res.det <- rbind(res.det, res.ext)
data.est <- rbind(res.det, res.ext)
data.est <- data.est[!duplicated(data.est[,!colnames(data.est)\%in\%c("sou", "value")]),]
data.est$ind <- factor(data.est$ind, levels = nameind)
data.est <- data.est[order(data.est$ind),]
data <- data.est

estimate(data=data.all, sources=est.array.d[est.array.d$est=="DET",], period=nameyear)
estimate(data=data.all, sources=est.array.d, period=nameyear)
}
\author{
OECD STAN
}
\seealso{
\code{\link{extend}}, \code{\link{detail}}
}
\keyword{estimate}

