% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example}
\alias{example}
\title{The data for the Quick-Start example in the vignette}
\description{
The data for the Quick-Start example in the vignette
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

