% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitHMM}
\alias{fitHMM}
\title{Fit a Hidden Markov Model}
\usage{
fitHMM(obs=list(), hmm, convergence=1e-06, maxIters=1000, dirFlags=list(), emissionProbs=list(), effectiveZero=0, verbose=FALSE, nCores=1, incrementalEM=FALSE, observationEmissionType=c())
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{hmm}{The initial Hidden Markov Model. This is a \code{\linkS4class{HMM}}.}

\item{convergence}{Convergence cutoff for EM-algorithm (default: 1e-6).}

\item{maxIters}{Maximum number of iterations.}

\item{dirFlags}{The flag sequence is needed when a bdHMM is fitted on undirected data (e.g.) ChIP only. It is a \code{list} of \code{character} vectors indication for each position its knwon directionality. U allows all states. F allows undirected states and states in forward direction. R allows undirected states and states in reverse direction.}

\item{emissionProbs}{List of precalculated emission probabilities of emission function is of type 'null'.}

\item{effectiveZero}{Transitions below this cutoff are analytically set to 0 to speed up comptuations.}

\item{verbose}{\code{logical} for printing algorithm status or not.}

\item{nCores}{Number of cores to use for computations.}

\item{incrementalEM}{When TRUE, the incremental EM is used to fit the model, where parameters are updated after each iteration over a single observation sequence.}

\item{observationEmissionType}{Only needed when HMMEmission is 'JointlyIndependent'. Defines for each dimension (columns in obs) of the data the type of emission to be used.}
}
\value{
A list containing the trace of the log-likelihood during EM learning and the fitted HMM model.
}
\description{
The function is used to fit (bidirectional) Hidden Markov Models, given one or more observation sequence.
}
\examples{
data(example)
hmm_fitted = fitHMM(observations, hmm_ex)
}
\seealso{
\code{\linkS4class{HMM}}
}

