% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{data}
\name{flags}
\alias{flags}
\title{Pre-computed flag sequence for the 'example' data.}
\description{
Pre-computed flag sequence for the 'example' data.
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\keyword{data}

