% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPosterior}
\alias{getPosterior}
\title{Calculate posterior state distribution.}
\usage{
getPosterior(hmm, obs=list(), emissionProbs=list(), dirFlags=list(), verbose=FALSE, nCores=1)
}
\arguments{
\item{hmm}{The initial Hidden Markov Model. This is a \code{\linkS4class{HMM}}.}

\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{emissionProbs}{List of precalculated emission probabilities of emission function is of type 'null'.}

\item{dirFlags}{The flag sequence is needed when a bdHMM is fitted on undirected data (e.g.) ChIP only. It is a \code{list} of \code{character} vectors indication for each position its knwon directionality. U allows all states. F allows undirected states and states in forward direction. R allows undirected states and states in reverse direction.}

\item{verbose}{\code{logical} for printing algorithm status or not.}

\item{nCores}{Number of cores to use for computations.}
}
\value{
A list containing for the observation sequences the posterior state (col) distribution at each position (row).
}
\description{
The function calculates posterior state probabilities for one or more observation sequence.
}
\examples{
data(example)
hmm_fitted = fitHMM(observations, hmm_ex)
posterior_hmm = getPosterior(hmm_fitted$hmm, observations)
}
\seealso{
\code{\linkS4class{HMM}}
}

