% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getViterbi}
\alias{getViterbi}
\title{Calculate the most likely state path}
\usage{
getViterbi(HMM, obs=list(), NAtol=5, emissionProbs=list(), verbose=FALSE)
}
\arguments{
\item{HMM}{The initial Hidden Markov Model. This is a \code{\linkS4class{HMM}}.}

\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{NAtol}{Successive positions having NAs longer than this threshold are masked in the viterbi path.}

\item{emissionProbs}{List of precalculated emission probabilities of emission function is of type 'null'.}

\item{verbose}{\code{logical} for printing algorithm status or not.}
}
\value{
A list containint the vterbi paths.
}
\description{
Given a Hidden Markov Model, the function calculates the most likely state path (viterbi) for one or more observation sequence.
}
\examples{
data(example)
hmm_fitted = fitHMM(observations, hmm_ex)
viterbi_hmm = getViterbi(hmm_fitted$hmm, observations)
}
\seealso{
\code{\linkS4class{HMM}}
}

