% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy}
\usage{
hierarchy(file = system.file("extdata", "loadDim_indi3agg.csv", package =
  "stan"), agg.exclude, parent = FALSE, order = FALSE,
  colnameprefix = NULL)
}
\arguments{
\item{file}{the aggregation matrix in csv format.}

\item{agg.exclude}{character vector with aggregates to be excluded from the hierarchy.}

\item{parent}{logical to list all parents of industry instead of members.}

\item{order}{logical to create ordered vector of aggregates and contained sectors}
}
\description{
Create STAN industry hierarchy
}
\details{
Function to create an industry hierarchy based on a aggregation matrix with industries in columns and aggregates in rows. Membership is indicated with a '1'.
}
\examples{
hierarchy(file=system.file("extdata", "loadDim_indi3agg.csv", package = "stan"))
}
\author{
OECD STAN
}
\seealso{
\code{\link{indAggregate}}, \code{\link{loadDim}}
}
\keyword{dimensions}

