% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HMM-class}
\alias{.HMM}
\alias{HMM-class}
\alias{[,HMM,ANY,ANY,ANY-method}
\title{This class is a generic container for Hidden Markov Models.}
\description{
This class is a generic container for Hidden Markov Models.
}
\section{Slots}{

\describe{
\item{\code{initProb}}{Initial state probabilities.}

\item{\code{transMat}}{Transition probabilities}

\item{\code{emission}}{Emission parameters as an HMMEmission object.}

\item{\code{nStates}}{Number of states.}

\item{\code{status}}{of the HMM. On of c('initial', 'EM').}

\item{\code{stateNames}}{State names.}

\item{\code{dimNames}}{Names of data tracks.}

\item{\code{LogLik}}{Log likelihood of a fitted HMM.}
}}
\section{Methods}{
  \describe{
    \item{[}{get elements from the HMM}
    }
}
\examples{

nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)

HMM(initProb=initProb, transMat=transMat, emission=HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means)), nStates=nStates, status='initial')
}
\seealso{
\code{\linkS4class{HMMEmission}}
}

