% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{HMM}
\alias{HMM}
\title{Create a HMM object}
\usage{
HMM(initProb = numeric(), transMat = matrix(numeric(), ncol = 1, nrow = 1),
  emission, nStates = numeric(), status = character(),
  stateNames = character(), dimNames = character(), LogLik = numeric())
}
\arguments{
\item{initProb}{Initial state probabilities.}

\item{transMat}{Transition probabilities}

\item{emission}{Emission parameters as an HMMEmission object.}

\item{nStates}{Number of states.}

\item{status}{of the HMM. On of c('initial', 'EM').}

\item{stateNames}{State names.}

\item{dimNames}{Names of data tracks.}

\item{LogLik}{Log likelihood of a fitted HMM.}
}
\description{
This function creates a HMM object.
}
\examples{
nStates = 5
means = list(4,11,4,11,-1)
Sigma = lapply(list(4,4,4,4,4), as.matrix)
transMat = matrix(1/nStates, nrow=nStates, ncol=nStates)
initProb = rep(1/nStates, nStates)
HMM(initProb=initProb, transMat=transMat, emission=HMMEmission(type='Gaussian', parameters=list(mu=means, cov=Sigma), nStates=length(means)), nStates=nStates, status='initial')

}
\seealso{
\code{\linkS4class{HMMEmission}}
}

