% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indAggregate.R
\name{indAggregate}
\alias{indAggregate}
\alias{listMissInd}
\title{Aggregate ISIC}
\usage{
indAggregate(data = stop("'data' must be specified"), isic = 3,
  cumulative = FALSE, naAsZero = NULL, fill2D = FALSE,
  missing.2d = NULL, part = FALSE, print.missing = FALSE)

listMissInd(data, isic, drop = FALSE)
}
\arguments{
\item{data}{a dataframe having STAN industry labels (ISIC Rev. 3 or ISIC Rev. 4) as column names.}

\item{isic}{an integer specifying the ISIC classification of data sources.}

\item{cumulative}{logical to cumulate values if aggregate already existing in data.}

\item{naAsZero}{character vector to specify industries where NAs are treated as zero, \code{ALL} for all industries.}

\item{fill2D}{logical add column with NA or zero for each missing 2-digit industry}

\item{missing.2d}{character vector with industries to impute zero, \code{NULL} for all 2-digit}

\item{part}{logical industry codes in column names contain partial information, e.g. \code{D01T03part}. Use with \code{fill2D=TRUE} in case of part aggregates and some 2-digits}
}
\description{
Aggregate industries ISIC Rev. 3 and ISIC Rev. 4

\code{listMissInd}: for each missing aggregate, list missing industries. Return NULL if all industries are missing for an aggregate
}
\details{
Calculate STAN ISIC Rev. 3 aggregates based on the A60 list and STAN ISIC Rev. 4 aggregates based on the A88 list.

listMissInd(data, isic)
}
\examples{
## ISIC Rev. 3:
x <- data.frame("cou"="JPN",
                "year"=c(rep(1995, 4), rep(1996,4)),
                "ind"=c("C01", "C02", "C05", "C10"),
                "value"=c(10, 10, 10, 10))
data <- dcast(x,cou + year ~ ind, value.var='value')
data.agg <- indAggregate(data=data, isic=3)
## ISIC Rev. 4:
x <- data.frame("cou"="JPN",
                "year"=c(rep(1995, 4), rep(1996,4)),
                "ind"=c("D01", "D02", "D03", "D05"),
                "value"=c(10, 10, 10, 10))
data <- dcast(x,cou + year ~ ind, value.var='value')
data.agg <- indAggregate(data=data, isic=4)
}
\author{
OECD STAN
}
\seealso{
\code{\link{loadDim}}
}
\keyword{dimensions}

