% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBdHMM.R
\name{initBdHMM}
\alias{initBdHMM}
\title{Initialization of bidirectional hidden Markov models}
\usage{
initBdHMM(obs, nStates, method, directedObs = rep(0, ncol(obs[[1]])), sizeFactors = matrix(1, nrow = length(obs), ncol = ncol(obs[[1]])), sharedCov = FALSE)
}
\arguments{
\item{obs}{The observations. A list of one or more entries containing the observation matrix (\code{numeric}) for the samples (e.g. chromosomes).}

\item{nStates}{The number of states.}

\item{method}{Emission distribution of the model. One out of c("NegativeBinomial", "PoissonLogNormal", "NegativeMultinomial", "ZINegativeBinomial", "Poisson", "Bernoulli", "Gaussian", "IndependentGaussian")}

\item{directedObs}{Integer vector defining the directionality (or strand-specificity) of the data tracks. Undirected (non-strand-specific) data tracks (e.g. ChIP) are indicated indicated by '0'. Directed (strand-specific) data tracks are indicated by increasing pairs of integers. For instance c(0,0,0,1,1,2,2): The first three data tracks are undirected, followed by two pairs of directed measurements.}

\item{sizeFactors}{Library size factors for Emissions PoissonLogNormal or NegativeBinomial as a length(obs) x ncol(obs[[1]]) matrix.}

\item{sharedCov}{If TRUE, (co-)variance of (Independent)Gaussian is shared over states. Only applicable to 'Gaussian' or 'IndependentGaussian' emissions. Default: FALSE.}
}
\value{
A HMM object.
}
\description{
Initialization of bidirectional hidden Markov models
}
\examples{

data(example)
hmm_ex = initHMM(observations, nStates=3, method="Gaussian") 

}

