% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{observations}
\alias{observations}
\title{Observation sequence of the Quick-Start example in the vignette}
\description{
Observation sequence of the Quick-Start example in the vignette
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

