% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{STAN-package}
\alias{STAN-package}
\title{STrand-specific ANnotation of genomic data}
\description{
STrand-specific ANnotation of genomic data
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\references{
B. Zacher, M. Lidschreiber, P. Cramer, J. Gagneur, and A. Tresch. Annotation of directed genomic states unveils variations in the Pol II transcription cycle. submitted, 2014.
}
\keyword{package}

