% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-package.R
\docType{package}
\name{STAN-package}
\alias{STAN-package}
\title{The genomic STate ANnotation package}
\description{
The genomic STate ANnotation package
}
\author{
Benedikt Zacher, Julia Ertl, Julien Gagneur, Achim Tresch
}
\references{
Zacher, B. and Lidschreiber, M. and Cramer, P. and Gagneur, J. and Tresch, A. (2014): Annotation of genomics data using bidirectional hidden Markov models unveils variations in Pol II transcription cycle Mol. Syst. Biol. 10:768
}
\keyword{package}

