\docType{package}
\name{stan-package}
\alias{stan}
\alias{stan-package}
\title{The stan Package}
\description{
  Structural Analysis.
}
\details{
  \tabular{ll}{ Package: \tab stan\cr Type: \tab Package\cr
  Version: \tab 0.1\cr Date: \tab 2013-09-30\cr Depends:
  \tab R (>= 2.10.0)\cr Encoding: \tab UTF-8\cr License:
  \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://www.oecd.org/sti/stan\cr }

  Provides functions to work with the OECD Structural
  Analysis Database
}
\author{
  STAN OECD \email{stan.contact@oecd.org}
}
\keyword{package}

