% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{yeastTF_databychrom_ex}
\alias{yeastTF_databychrom_ex}
\title{Processed ChIP-on-chip data for yeast TF example}
\description{
Processed ChIP-on-chip data for yeast TF example
}
\author{
Benedikt Zacher, Julien Gagneur, Achim Tresch
}
\keyword{data}

